+++ tests/test_baseflumine.py
@@ -441,7 +441,9 @@ class BaseFlumineTest(unittest.TestCase):
         mock_strategy = mock.Mock()
         self.base_flumine.strategies = [mock_strategy]
         mock_current_orders = mock.Mock(orders=[mock_order])
-        mock_event = mock.Mock(event=[mock_current_orders], venue=VenueType.BETFAIR)
+        mock_event = mock.Mock(
+            event=[mock_current_orders], venue=VenueType.BETFAIR, callback=None
+        )
         self.base_flumine._process_current_orders(mock_event)
         mock_process_current_orders.assert_called_with(
             self.base_flumine.markets,
@@ -470,7 +472,9 @@ class BaseFlumineTest(unittest.TestCase):
         mock_strategy = mock.Mock()
         self.base_flumine.strategies = [mock_strategy]
         mock_current_orders = mock.Mock(orders=[mock_order])
-        mock_event = mock.Mock(event=[mock_current_orders], venue=VenueType.BETDAQ)
+        mock_event = mock.Mock(
+            event=[mock_current_orders], venue=VenueType.BETDAQ, callback=None
+        )
         self.base_flumine._process_current_orders(mock_event)
         mock_process_betdaq_current_orders.assert_called_with(
             self.base_flumine.markets,
@@ -489,6 +493,35 @@ class BaseFlumineTest(unittest.TestCase):
         mock_event = mock.Mock(event=[])
         self.base_flumine._process_current_orders(mock_event)
         mock_process_current_orders.assert_not_called()
+
+    @mock.patch("flumine.baseflumine.utils.call_process_orders_error_handling")
+    def test__process_current_orders_callback(
+        self, mock_call_process_orders_error_handling
+    ):
+        mock_order = mock.Mock(complete=True)
+        mock_market = mock.Mock(closed=False)
+        mock_market.blotter.active = True
+        mock_market.blotter.strategy_orders.return_value = [mock_order]
+        self.base_flumine.markets = [mock_market]
+        mock_strategy = mock.Mock()
+        self.base_flumine.strategies = [mock_strategy]
+        mock_current_orders = mock.Mock(orders=[mock_order])
+        mock_callback = mock.Mock()
+        mock_event = mock.Mock(
+            event=[mock_current_orders], venue=VenueType.BETFAIR, callback=mock_callback
+        )
+        self.base_flumine._process_current_orders(mock_event)
+        mock_callback.assert_called_with(
+            self.base_flumine.markets,
+            self.base_flumine.strategies,
+            mock_event,
+            self.base_flumine.log_control,
+            self.base_flumine._add_market,
+        )
+        mock_market.blotter.strategy_orders.assert_called_with(mock_strategy)
+        mock_call_process_orders_error_handling.assert_called_with(
+            mock_strategy, mock_market, [mock_order]
+        )
 
     def test__process_custom_event(self):
         mock_market = mock.Mock()